Python/Frameworks/Django/Gymkhana Stampede/requirements.txt
absl-py==0.3.0
astor==0.7.1
attrs==18.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
botocore==1.10.65
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
colorama==0.3.9
credentials==1.1
dask==0.18.2
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
Django==2.1rc1
django-allauth==0.36.0
django-allauth-bootstrap==0.2
django-bootstrap==0.2.4
django-bootstrap-form==3.4
django-cities==0.5.0.6
django-geoposition==0.3.0
django-googlecloud-storage==0.0.1
django-localflavor==2.0
django-model-utils==3.1.2
docutils==0.14
easyapiio==0.1.31
entrypoints==0.2.3
Flask==1.0.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.90.4
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.0
google==2.0.1
google-api-core==1.2.1
google-auth==1.5.0
google-cloud==0.33.1
google-cloud-bigquery==1.4.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.29.0
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
google-cloud-dns==0.29.0
google-cloud-error-reporting==0.30.0
google-cloud-firestore==0.29.0
google-cloud-language==1.0.2
google-cloud-logging==1.6.0
google-cloud-monitoring==0.30.1
google-cloud-pubsub==0.35.4
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==1.4.0
google-cloud-speech==0.35.0
google-cloud-storage==1.10.0
google-cloud-trace==0.19.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.3.0
google-cloud-vision==0.33.0
google-gax==0.16.0
google-resumable-media==0.3.1
googleapis-common-protos==1.6.0b4
grpc-google-iam-v1==0.11.4
grpcio==1.14.0rc1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
ImageScraper==2.0.7
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.3.1
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==3.0.0a1
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.2.3
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.15.0
oauth2client==4.1.2
oauthlib==2.1.0
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
ply==3.11
prometheus-client==0.3.0
prompt-toolkit==1.0.15
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.0
psutil==5.4.6
pur==5.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyreadline==2.1
pyrsistent==0.14.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyWavelets==0.5.2
pywinpty==0.5.4
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
SimplePool==0.1
six==1.11.0
swapper==1.1.0
tensorboard==1.10.0
tensorflow==1.10.0rc0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.24.0
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.3.1
win-unicode-console==0.5
